24/07/21
Completed the health UI, with adjustable hearts. Also have implimented a soft collision for the enemies (rabbits) so that they don't overlap and cant bunch up to avoid looking like one single entity.
Have added trees, but will need to change the color in photoshop.
Next, we are working on the Player camera and how it will follow the player.

KNOWN ISSUES: The player bugs when walking around the bush or trees, sticking to the side.
The animation for a hit (the hit effect) doesn't disappear or play, just spawn and stays there.